+++ lib/core/constants/api_end_point.dart
@@ -1,0 +1,4 @@
+class ApiEndPoint {
+  static const String refresh = '/auth/refresh';
+  static const String login = '/auth/login';
+}