+++ lib/core/router/app_router.dart
@@ -1,4 +1,5 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:sample_project/features/login/presentation/presentation.dart';
 import 'package:sample_project/features/splash/screen/splash_screen.dart';
 
 part 'app_router.gr.dart';
@@ -7,6 +8,7 @@ part 'app_router.gr.dart';
 class AppRouter extends RootStackRouter {
   @override
   List<AutoRoute> get routes => [
-    AutoRoute(initial: true, page: SplashRoute.page),
+    AutoRoute(page: SplashRoute.page),
+    AutoRoute(initial: true, page: LoginRoute.page),
   ];
 }