+++ lib/features/login/data/models/mapping/token_entity_mapping.dart
@@ -1,0 +1,7 @@
+import 'package:sample_project/features/login/data/data.dart';
+import 'package:sample_project/features/login/domain/domain.dart';
+
+extension TokenEntityMapping on TokenModel {
+  TokenEntity toEntity() =>
+      TokenEntity(accessToken: accessToken, refreshToken: refreshToken);
+}