+++ lib/features/login/domain/entity/login_request_entity.dart
@@ -1,0 +1,5 @@
+class LoginRequestEntity {
+  final String username;
+  final String password;
+  const LoginRequestEntity({required this.username, required this.password});
+}