+++ lib/features/login/presentation/bloc/login_event.dart
@@ -1,0 +1,19 @@
+part of 'login_bloc.dart';
+
+sealed class LoginEvent {
+  const LoginEvent();
+}
+
+final class UserNameChangedEvent extends LoginEvent {
+  final String username;
+  UserNameChangedEvent(this.username);
+}
+
+final class PasswordChangedEvent extends LoginEvent {
+  final String password;
+  PasswordChangedEvent(this.password);
+}
+
+final class LoginSubmittedEvent extends LoginEvent {
+  const LoginSubmittedEvent();
+}