+++ lib/main.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sample_project/core/core.dart';
+import 'package:sample_project/features/login/data/repository/login_repository_impl.dart';
+import 'package:sample_project/features/login/presentation/bloc/login_bloc.dart';
 
 void main() => runApp(const SampleProjectApp());
 
@@ -9,9 +12,17 @@ class SampleProjectApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final router = AppRouter();
-    return MaterialApp.router(
-      title: AppConstant.appName,
-      routerConfig: router.config(),
+    return MultiBlocProvider(
+      providers: [
+        BlocProvider(
+          create: (_) =>
+              LoginBloc(loginRepository: LoginRepositoryImpl(NetworkService())),
+        ),
+      ],
+      child: MaterialApp.router(
+        title: AppConstant.appName,
+        routerConfig: router.config(),
+      ),
     );
   }
 }