+++ changelog.md
@@ -1,4 +1,11 @@
 @sushantebp
+
 # Project Setup (develop)
+
 - 14 oct
 - setup with autoroute and custom widgets
+
+# feature/login
+
+- implemented network service (dio) with interceptors ,secure storage service (flutter_secure_storage)
+- implemented converting dart object,models and mapping them